chore: separate the common parts
